Rename Parse namespace to CodeGen
