+++ src/CodeGen/ParsedMethod.php
@@ -14,10 +14,24 @@ declare(strict_types=1);
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Bake\Parse;
+namespace Bake\CodeGen;
 
-use Cake\Core\Exception\CakeException;
-
-class ParseException extends CakeException
+/**
+ * @internal
+ */
+final class ParsedMethod
 {
+    /**
+     * @param string $name name
+     * @param string $code Code block
+     * @param string|null $docblock Docblock
+     * @param array $attributes Attributes
+     */
+    public function __construct(
+        public readonly string $name,
+        public readonly string $code,
+        public readonly ?string $docblock,
+        public readonly array $attributes
+    ) {
+    }
 }
